Rename NamespaceCreate to NamespaceEdit
