+++ src/componentsNext/NamespaceEdit/InfoTooltip.tsx
