+++ src/componentsNext/NamespaceEdit/index.tsx
