+++ src/pages/OnboardingPage.tsx
@@ -7,7 +7,7 @@ import Alert from "~/design/Alert";
 import ApiKeyPanel from "./namespace/Settings/ApiKey";
 import Button from "~/design/Button";
 import Logo from "~/design/Logo";
-import NamespaceCreate from "~/componentsNext/NamespaceCreate";
+import NamespaceCreate from "~/componentsNext/NamespaceEdit";
 import { pages } from "~/util/router/pages";
 import { useListNamespaces } from "~/api/namespaces/query/get";
 import { useNavigate } from "react-router-dom";
